Little changes to import / export to handle custom charts
